input에서 하던 예외처리 이동 - 중복체크 > LottoMachine, LottoWinningNumber - 로또숫자개수 > Lotto - 테스트 수정
